requirements.txt
fastapi
jinja2
markdown2
python-dotenv
requests
uvicorn